requirements.txt
streamlit==1.31.1
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.4.0
python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.31.0